+++ src/requisition-search/requisition-search.controller.js
@@ -133,6 +133,11 @@
          */
         vm.offline = undefined;
 
+        vm.options = {
+            'createdDate,desc': 'requisitionSearch.date',
+            'status': 'requisitionSearch.status'
+        };
+
         /**
          * @ngdoc method
          * @methodOf requisition-search.controller:RequisitionViewController